pyproject.toml
[tool.poetry]
name = "cached-historical-data-fetcher"
version = "0.1.0"
description = "Python utility for fetching any historical data using caching."
authors = ["34j <34j.95a2p@simplelogin.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/34j/cached-historical-data-fetcher"
documentation = "https://cached-historical-data-fetcher.readthedocs.io"
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries",
]
packages = [
    { include = "cached_historical_data_fetcher", from = "src" },
]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/34j/cached-historical-data-fetcher/issues"
"Changelog" = "https://github.com/34j/cached-historical-data-fetcher/blob/main/CHANGELOG.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
numpy = "^1.26.0"
pandas = "^2.1.1"
joblib = "^1.3.2"
aiofiles = "^23.2.1"
tqdm = "^4.66.1"
python-slugify = "^8.0.1"
lz4 = "^4.3.2"
typing-extensions = "^4.8.0"

[tool.poetry.group.dev.dependencies]
pre-commit = ">=3"
pytest = "^7.0"
pytest-cov = "^3.0"
parameterized = "^0.9.0"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
myst-parser = ">=0.16"
sphinx = ">=4.0"
sphinx-rtd-theme = ">=1.0"


[tool.poetry.group.notebook.dependencies]
ipykernel = "^6.25.2"
aiohttp = "^3.8.5"
lxml = "^4.9.3"
beautifulsoup4 = "^4.12.2"
jupyter = "^1.0.0"
ipywidgets = "^8.1.1"
aiohttp-client-cache = {extras = ["all"], version = "^0.9.1"}

[tool.semantic_release]
branch = "main"
version_toml = ["pyproject.toml:tool.poetry.version"]
version_variable = "src/cached_historical_data_fetcher/__init__.py:__version__"
build_command = "pip install poetry && poetry build"

[tool.semantic_release.changelog]
exclude_commit_patterns = [
    "chore*",
    "ci*",
]

[tool.semantic_release.changelog.environment]
keep_trailing_newline = true

[tool.pytest.ini_options]
addopts = "-v -Wdefault --cov=cached_historical_data_fetcher --cov-report=term-missing:skip-covered"
pythonpath = ["src"]

[tool.coverage.run]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "@overload",
    "if TYPE_CHECKING",
    "raise NotImplementedError",
    'if __name__ == "__main__":',
]

[tool.isort]
profile = "black"
known_first_party = ["cached_historical_data_fetcher", "tests"]

[tool.autoflake]
expand_star_imports = true
remove_all_unused_imports = true

[tool.mypy]
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
mypy_path = "src/"
no_implicit_optional = true
show_error_codes = true
warn_unreachable = true
warn_unused_ignores = true
exclude = [
    'docs/.*',
    'setup.py',
]

[[tool.mypy.overrides]]
module = "tests.*"
allow_untyped_defs = true

[[tool.mypy.overrides]]
module = "docs.*"
ignore_errors = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
